backend/auth/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
asyncpg
passlib[bcrypt]
pydantic[email]
